+++ packages/graph-explorer/src/components/DefaultQueryErrorBoundary.tsx
@@ -1,0 +1,21 @@
+import { QueryErrorResetBoundary } from "@tanstack/react-query";
+import { PropsWithChildren } from "react";
+import { ErrorBoundary } from "react-error-boundary";
+import PanelError from "./PanelError";
+
+export function DefaultQueryErrorBoundary({ children }: PropsWithChildren) {
+  return (
+    <QueryErrorResetBoundary>
+      {({ reset }) => (
+        <ErrorBoundary
+          onReset={reset}
+          fallbackRender={({ resetErrorBoundary, error }) => (
+            <PanelError error={error} onRetry={resetErrorBoundary} />
+          )}
+        >
+          {children}
+        </ErrorBoundary>
+      )}
+    </QueryErrorResetBoundary>
+  );
+}