+++ Changelog.md
@@ -4,6 +4,8 @@
 
 ### Other changes
 
+- **Added** ability to customize default neighbor expansion limit
+  ([#925](https://github.com/aws/graph-explorer/pull/925))
 - **Improved** Improved connection details by moving the sync button near the
   last sync timestamp (thanks @enumura1,
   [#901](https://github.com/aws/graph-explorer/pull/901),